Baltimore City Public Schools Enrollment and Demographic data (SY 2021-2022)

Total enrollment and demographic characteristics for Baltimore City public school students including data aggregated by grade, grade band, school, and citywide for the 2021-2022 school year.

Usage

enrollment_demographics_SY2122

Format

A data frame with 2,153 rows and 20 variables:

school_number

School number

school_name

School name

management_type

School management type

grade_band

Grade band (factor)

grade

Grade (factor)

grade_range

Grade range (factor)

total_enrollment

Total Enrollment

percent_male

Percent Male

percent_female

Percent Female

percent_nonbinary

Percent Nonbinary

percent_direct_certification

Percent Direct Certification

percent_swd

Percent Students with Disabilities (SWD)

percent_el

Percent English Learner (EL)

percent_non_hispanic_black

Percent Non-Hispanic Black

percent_non_hispanic_white

Percent Non-Hispanic White

percent_hispanic

Percent Hispanic

percent_non_hispanic_asian

Percent Non-Hispanic Asian

percent_non_hispanic_american_indian

Percent Non-Hispanic American Indian

percent_non_hispanic_pacific_islander

Percent Non-Hispanic Native Hawaiian or Other Pacific Islander

percent_non_hispanic_multiracial

Percent Non-Hispanic Multiracial

Source

https://www.baltimorecityschools.org/data

Details

The column percent_non_hispanic_pacific_islander is named Percent non-Hispanic Hawaiian and Other Pacific Islander in the source but is renamed to match the prior two years of data. This data also includes a `percent_nonbinary column that is not present in prior years.

For variables with percentages, a value of 0.05 may represent <=5% rather than 5% and a value of 0.95 may represent >=95% rather than 95%.