Enrolled students by school and grade (SY 2009-2019)

Enrollment data from the Maryland State Department of Education (MSDE). Raw CSV data files with the full state enrollment data are available in the inst/extdata folder. The number of enrolled students includes ungraded special education and pre-kindergarten students. Variable definitions are based on the definitions from the MSDE website.

enrollment_msde_SY0919

Format

A data frame with 16,763 rows and 6 variables:

school_year

School or academic year for enrollment count, e.g. 2019 data is from the start of the 2019-2020 school year.

school_number

School number (0 indicates all schools)

school_name

School name

enrolled_count

Number of students registered to attend the school as of September 30 in the grade or grade range

grade

Grades from Prekindergarden (PK) to Grade 12

grade_range

Elementary school, middle school, high school, or all grades (total enrollment)

Source

MSDE Data Downloads