Baltimore City Public School Parent Survey Results (SY 2018-2019) - long format

Results from the Baltimore City Public School Parent Survey administered during the for the 2018-2029 school year. Data includes composite scores and responses to selected questions school. Long format version is based on the parent_survey_SY1819 data with an additional label column based on the original Excel sheet column names.

parent_survey_SY1819_long

Format

A data frame with 7,224 rows and 10 variables:

school_number

School number

school_name

School name

management_type

School management type

grade_band

School grade band

parent_response_count

Number of parent survey responses

est_num_parents

Estimated number of eligible parents

est_response_rate

Estimated parent response rate

variable

Unique slug for composite score or question

value

Value of composite score or share of parents responding affirmatively to question

label

Variable label for composite score or question

Source

https://www.baltimorecityschools.org/data