Attendance data by school (SY 2009-2019)

Attendance data from the Maryland State Department of Education (MSDE). Raw CSV data files with the full state enrollment data are available in the inst/extdata folder. Any percent variables with a value of 95 percent should be interpreted as equivalent at or above 95 percent. Any percent variables with a value of 5 percent should be interpreted as equivalent to at or below 5 percent. Chronic absenteeism data is not available prior to 2018 when the federal Every Student Succeeds Act (passed in 2015) went into effect and Maryland began to include the chronic absenteeism rate in the state's annual academic report cards. Variable definitions are based on the definitions from the MSDE website.

Usage

attendance_msde_SY0919

Format

A data frame with 3,197 rows and 15 variables:

school_year

School year or academic year, e.g. 2019 for 2019-2020 school year

school_number

School number (0 indicates all schools)

school_name

School name

grade_band

Grade band (school type)

attend_rate_pct

Attendance rate as defined as the average daily attendance rate of students in the school and grade band, including special education students

days_attended_cnt

Total number of students in attendance from the first day of school through April 13th

days_member_cnt

Total number of students registered to attend from the first day of school through April 13th

fewer_5_pct

Percent of students absentee rate is the percentage of students absent from school for five or fewer days between the first day of the school year and the last day of the school year

fewer_5_cnt

Count of students absentee rate is the percentage of students absent from school for five or fewer days between the first day of the school year and the last day of the school year

more_20_pct

Percent of students absentee rate is the percentage of students absent from school for twenty or more days between the first day of the school year and the last day of the school year

more_20_cnt

Number of students absentee rate is the percentage of students absent from school for twenty or more days between the first day of the school year and the last day of the school year

more_90_member_cnt

Number of students with an attendance rate of 90 percent or more (?)

chronic_absentee_cnt

Number of students expected to attend school for at least 10 days who were absent 10 percent or more of the school days while enrolled at that school

chronic_absentee_denom

Number of students who are expected to attend school for at least 10 days in the school year

chronic_absentee_pct

Chronic absenteeism rate as defined as the percent of students expected to attend school for at least 10 days who were absent 10 percent or more of the school days while enrolled at that school

Source

MSDE Data Downloads