MSDE Kindergarten Readiness Assessment (KRA) results for Baltimore City Public Schools (SY 2019-2020)

As of 2018-19, students who are not scored (NS) in all items still receive a composite score and readiness level. NS is applied when an item requires demonstration of a skill (e.g., hopping), and the child is not able to respond due to a disability that restricts or prevents participating in the skill being assessed. A NS rating would not be appropriate when the response to the item reflects a child's functioning at an earlier developmental level and their ability to respond is not affected by a disability; in this instance the appropriate rating would be “0” since the child could access the item, but did not demonstrate the skill. See MSDE: KRA 2018-19 Technical Report, January 2019. Consistent with MSDE's reporting, aggregates exclude students with some incomplete items, and include students with not scorable (NS) items. Missing values include data that has been suppressed due to a group size of less than 10 students.

Children whose readiness skills and behaviors are "approaching" or "emerging" overall require instructional support to be successful in kindergarten and beyond. Reports of KRA data prior to the 2018-19 administration have readiness level assigned for each of the four domains (Literacy, Math, Social Foundations, and Physical Development). Since last year's administration, readiness levels are no longer reported for individual domains, and only an overall or composite readiness level is reported. This is why this data file does not resemble previous releases of KRA data. Details on the KRA assessment and changes to the test can be found at https://pd.kready.org/r4kmaryland.

Usage

kra_results_SY1920

Format

A data frame with 1604 rows and 13 variables:

school_number

School number (0 indicates all schools)

school_name

School name

student_group

Student group (includes "All Students")

group_type

Student group type

num_tested_scored_overall

Number of test takers with overall scores

avg_scale_score

Average KRA scale score

perc_emerging_readiness

Percent of test takers in Emerging Readiness category. Emerging readiness: Student demonstrates limited foundational skills and behaviors that prepare him or her for curriculum based on Maryland College and Career-Ready kindergarten standards.

perc_approaching_readiness

Percent of test takers in Approaching Readiness category. Approaching readiness: Student demonstrates some foundational skills and behaviors that prepare him or her for curriculum based on Maryland College and Career-Ready kindergarten standards.

perc_demonstrating_readiness

Percent of test takers in Demonstrating Readiness category. Demonstrating readiness: Student demonstrates foundational skills and behaviors that prepare him or her for curriculum based on Maryland College and Career-Ready kindergarten standards.

avg_language_and_literacy_score

Average score in Literacy domain

avg_math_score

Average score in KRA Math domain

avg_physical_development_score

Average score in Physical Development domain

avg_social_foundations_score

Average score in Social Foundations domain

Source

BCPSS Public 2019-20 KRA School Level Results